hit8/requirements.txt
selenium==4.21.0
pytest==8.2.2
pytest-cov==5.0.0
pre-commit
psycopg2-binary
pandas
tabulate
undetected-chromedriver